refactor: update web app components filenames
